fix: fix outbox repository pattern
