+++ services/geospatial/src/Geospatial.Application/UseCases/IntersectsUseCase.cs
@@ -1,7 +1,7 @@
-using Geospatial.Application.Interfaces.Repositories;
 using Geospatial.Application.Interfaces.UseCases;
-using Geospatial.Application.Models;
+using Geospatial.Domain.Events;
 using Geospatial.Domain.Geometry;
+using Geospatial.Domain.Repositories;
 using Geospatial.Domain.Services;
 using Microsoft.Extensions.Logging;
