+++ services/geospatial/src/Geospatial.Domain/Events/GeoLocation.cs
@@ -1,0 +1,7 @@
+namespace Geospatial.Domain.Events;
+
+public class GeoLocation
+{
+    public double Lat { get; set; }
+    public double Lon { get; set; }
+}
