--- services/geospatial/src/Geospatial.Domain/Repositories/IOutboxRepository.cs
@@ -1,7 +1,0 @@
-﻿namespace Geospatial.Domain.Repositories;
-
-public interface IOutboxRepository
-{
-    public Task<bool> ExistsPending(Guid aggregateId);
-    public Task<bool> UpdateProcessed(Guid id);
-}
