+++ services/geospatial/tests/Geospatial.UnitTests/UseCases/ContainsPointUseCaseTests.cs
@@ -1,6 +1,7 @@
-using Geospatial.Application.Interfaces.Repositories;
 using Geospatial.Application.UseCases;
+using Geospatial.Domain.Events;
 using Geospatial.Domain.Geometry;
+using Geospatial.Domain.Repositories;
 using Geospatial.Domain.Services;
 using Geospatial.Domain.ValueObjects;
 using Microsoft.Extensions.Logging;
@@ -19,7 +20,7 @@ public class ContainsPointUseCaseTests
     {
         _mockCalculator = new Mock<IGeospatialCalculator>();
         _mockEventRepository = new Mock<IGeospatialEventRepository>();
-        _mockEventRepository.Setup(x => x.IndexAsync(It.IsAny<Geospatial.Application.Models.GeospatialOperationEvent>(), It.IsAny<CancellationToken>()))
+        _mockEventRepository.Setup(x => x.IndexAsync(It.IsAny<GeospatialOperationEvent>(), It.IsAny<CancellationToken>()))
             .Returns(Task.CompletedTask);
         _mockLogger = new Mock<ILogger<ContainsPointUseCase>>();
         _sut = new ContainsPointUseCase(_mockCalculator.Object, _mockEventRepository.Object, _mockLogger.Object);
@@ -37,7 +38,7 @@ public class ContainsPointUseCaseTests
 
         Assert.True(result);
         _mockCalculator.Verify(x => x.ContainsPoint(area, point), Times.Once);
-        _mockEventRepository.Verify(x => x.IndexAsync(It.IsAny<Geospatial.Application.Models.GeospatialOperationEvent>(), It.IsAny<CancellationToken>()), Times.Once);
+        _mockEventRepository.Verify(x => x.IndexAsync(It.IsAny<GeospatialOperationEvent>(), It.IsAny<CancellationToken>()), Times.Once);
     }
 
     [Fact]
@@ -52,7 +53,7 @@ public class ContainsPointUseCaseTests
 
         Assert.False(result);
         _mockCalculator.Verify(x => x.ContainsPoint(area, point), Times.Once);
-        _mockEventRepository.Verify(x => x.IndexAsync(It.IsAny<Geospatial.Application.Models.GeospatialOperationEvent>(), It.IsAny<CancellationToken>()), Times.Once);
+        _mockEventRepository.Verify(x => x.IndexAsync(It.IsAny<GeospatialOperationEvent>(), It.IsAny<CancellationToken>()), Times.Once);
     }
 
     private static GeoPolygon CreateTestPolygon()
