+++ services/geospatial/tests/Geospatial.UnitTests/UseCases/IntersectsUseCaseTests.cs
@@ -1,6 +1,7 @@
-using Geospatial.Application.Interfaces.Repositories;
 using Geospatial.Application.UseCases;
+using Geospatial.Domain.Events;
 using Geospatial.Domain.Geometry;
+using Geospatial.Domain.Repositories;
 using Geospatial.Domain.Services;
 using Geospatial.Domain.ValueObjects;
 using Microsoft.Extensions.Logging;
@@ -19,7 +20,7 @@ public class IntersectsUseCaseTests
     {
         _mockCalculator = new Mock<IGeospatialCalculator>();
         _mockEventRepository = new Mock<IGeospatialEventRepository>();
-        _mockEventRepository.Setup(x => x.IndexAsync(It.IsAny<Geospatial.Application.Models.GeospatialOperationEvent>(), It.IsAny<CancellationToken>()))
+        _mockEventRepository.Setup(x => x.IndexAsync(It.IsAny<GeospatialOperationEvent>(), It.IsAny<CancellationToken>()))
             .Returns(Task.CompletedTask);
         _mockLogger = new Mock<ILogger<IntersectsUseCase>>();
         _sut = new IntersectsUseCase(_mockCalculator.Object, _mockEventRepository.Object, _mockLogger.Object);
