+++ services/ingestion/src/Ingestion.Api/appsettings.json
@@ -6,7 +6,37 @@
     }
   },
   "AllowedHosts": "*",
+  "ConnectionStrings": {
+    "IngestionReadDatabase": "mongodb://mongo:27017/?authSource=admin",
+    "IngestionWriteDatabase": "Host=postgres;Port=5432;Database=meuBanco;Username=yeti;Password=1234;",
+    "Kafka": "kafka:9092",
+    "Redis": "redis:6379"
+  },
+  "MinIO": {
+    "Host": "minio",
+    "Port": "9000",
+    "AccessKey": "minioadmin",
+    "SecretKey": "minioadmin"
+  },
+  "ElasticSearch": {
+    "URI": "http://elasticsearch:9200"
+  },
   "Geospatial": {
-    "BaseUrl": "http://localhost:5170"
+    "BaseUrl": "http://geospatial:8080"
+  },
+  "Polling": {
+    "Fire": {
+      "BaseUrl": "https://firms.modaps.eosdis.nasa.gov",
+      "DataSourceId": "11111111-1111-1111-1111-111111111111",
+      "TenantId": "22222222-2222-2222-2222-222222222222",
+      "Area": "BRA",
+      "Days": "30",
+      "Date": "2025-08-01"
+    },
+    "Earthquake": {
+      "BaseUrl": "https://earthquake.usgs.gov",
+      "DataSourceId": "33333333-3333-3333-3333-333333333333",
+      "TenantId": "44444444-4444-4444-4444-444444444444"
+    }
   }
 }
