+++ services/ingestion/src/Ingestion.Infrastructure.Write/Persistence/Repositories/OutboxRepository.cs
@@ -25,7 +25,16 @@ public class OutboxRepository : IOutboxRepository
         return outboxMessage.Id;
     }
 
-    public async Task<IEnumerable<OutboxRow>> GetPending() =>
+    public async Task<IEnumerable<OutboxRow>> GetPendingAsync() =>
         await _writeDbContext.Connection.QueryAsync<OutboxRow>(
             "SELECT id, outbox_type, payload FROM outbox WHERE processed = false LIMIT 50 FOR UPDATE SKIP LOCKED");
+    
+    public async Task<bool> UpdateProcessedAsync(Guid id)
+    {
+        var affectedRows = await _writeDbContext.Connection.ExecuteAsync(
+            "UPDATE outbox SET processed = true, processed_at = now() WHERE aggregate_id  = @AggregateId",
+            new { AggregateId = id });
+
+        return affectedRows > 0;
+    }
 }
