+++ services/risk-catalog/src/RiskCatalog.Api/appsettings.json
@@ -6,7 +6,23 @@
     }
   },
   "AllowedHosts": "*",
+  "ConnectionStrings": {
+    "RiskCatalogDatabase": "Host=postgres;Port=5432;Database=risk_catalog;Username=postgres;Password=postgres",
+    "Redis": "redis:6379",
+    "Kafka": "kafka:9092"
+  },
+  "ElasticSearch": {
+    "URI": "http://elasticsearch:9200"
+  },
   "Geospatial": {
-    "BaseUrl": "http://localhost:5170"
+    "BaseUrl": "http://geospatial:8080"
+  },
+  "Jwt": {
+    "Issuer": "sentinel",
+    "Audience": "sentinel",
+    "Key": "super-secret-key-change-this"
+  },
+  "Scalar": {
+    "AuthToken": "changeme"
   }
 }
